app/pyproject.toml
[project]
name = "herta-agent"
version = "0.0.1"
dependencies = [
    "langgraph>=0.2.6",
    "langchain>=0.3",
    "langchain-community>=0.2.17",
    "langchain-openai>=0.3.2",
    "requests>=2.32.3",
    "requests_cache>=1.2.1",
    "python-dotenv>=1.0.1",
]

